requirements.txt
numpy == 1.18.4
pandas == 1.2.4
PyYAML == 5.4.1 
